adding main file to project for option
