--- rem.md
@@ -1,68 +1,0 @@
-# HTMLForms
-+ Create a blank HTML document.
-+ Plan out how the elements on the page need to be grouped together.
-+ Lay out the empty semantic HTML elements for the major sections (form, sidebar, footer…) and make them the proper size and + alignment. It may be helpful to change their background colors while you’re positioning them on the page. Don’t forget to use 
-+ your browser’s developer tools (right click on the page, click “inspect element”)!
-+ Fill in those elements with the content you want to include.
-+ Now style them to look like the real webpage.
-+ Push your solution up to Github and you’re good to go!
-
-# SignUp Form from mint.com
-
-> This project consists of building an HTML document that matches the appearance of mint.com’s signup page
-
-You can find the original project specification at: https://www.theodinproject.com/courses/html5-and-css3/lessons/html-forms
-
-![screenshot](./img/screem-site.png)
-
-## Built With
-
-- CSS and HTML
-- Flexbox
-- Form and input tags
-- Media queries
-- Fontawesome
-
-## Live Demo
-
-[Live Demo Link](https://rawcdn.githack.com/enelesmai/HTMLForms/71d392b1e96e38074c4e29a6c84f5d15fcab9680/index.html)
-
-
-## Getting Started
-
-**This is a static page and was made using  CSS & HTML, you don't need .**
-**anything to start up this page, is already published in a public repository so you can check the final result. For a better understanding of the code read about FlexBox.**
-
-
-## Authors
-
-👤 **Xóchitl Selene Flores Pérez**
-
-- Github: [@enelesmai](https://github.com/enelesmai)
-- Twitter: [@enelesmai](https://twitter.com/enelesmai)
-- Linkedin: [xochitlselene](https://linkedin.com/in/xochitlselene)
-
-👤 **Ruben Paz Chuspe**
-
-- Github: [@rubenpazch](https://github.com/rubenpazch)
-- Linkedin: [rubenpch](https://www.linkedin.com/in/rubenpch/)
-
-## 🤝 Contributing
-
-Contributions, issues and feature requests are welcome!
-
-Feel free to check the (https://github.com/enelesmai/htmlforms/issues).
-
-## Show your support
-
-Give a ⭐️ if you like this project!
-
-## Acknowledgments
-
-- Hat tip to anyone whose code was used
-- Inspiration
-- etc
-
-## 📝 License
-
-This project is [MIT](lic.url) licensed.
